Push blobs before pushing manifests
We need to push blobs to the registry before pushing manifests, else the
`oras manifest push` command will fail with 'unknown blob' error. The
reason is that the manifest references blobs by their digests, which
means that if you push the manifest but somehow don't push the blob,
you'll end up with dangling references. Some registries (like GHCR)
prohibit this behavior and fail the push action.