Push the config blob alongside the signature blob
In previous versions, the "config blob" was not stored and pushed,
leading to `cosign save` not working, because it was encountering
inexistent digests.

We didn't find out about this issue earlier, because when testing with
the `dangerzone-testing` images, the CI was publishing signatures (using
a different key) and the configuration blobs, which are the same.